ImageGeneration/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py39h06a4308_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.6=pypi_0
dm-tree=0.1.8=pypi_0
etils=1.0.0=pypi_0
flatbuffers=23.1.21=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.58.0=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.8.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
jax=0.4.3=pypi_0
jaxlib=0.4.3+cuda11.cudnn86=pypi_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=15.0.6.1=pypi_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.3=pypi_0
ml-collections=0.1.1=pypi_0
ncurses=6.4=h6a678d5_0
ninja=1.11.1=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py39h06a4308_0
promise=2.3=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.0=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.40.1=h5082296_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-datasets=4.8.2=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-gan=2.1.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.30.0=pypi_0
tensorflow-metadata=1.12.0=pypi_0
tensorflow-probability=0.19.0=pypi_0
termcolor=2.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.10=h5eee18b_1
zipp=3.13.0=pypi_0
zlib=1.2.13=h5eee18b_0
